requirements.txt
# Core dependencies - using flexible versions for compatibility
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
torchaudio>=2.0.0,<3.0.0

# Diffusion models
diffusers>=0.21.0
transformers>=4.30.0

# Image processing
Pillow>=9.5.0
numpy>=1.24.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# HTTP client
requests>=2.31.0

# Environment configuration
python-dotenv>=1.0.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.1.0

# Optional: for better Stable Diffusion performance
# Uncomment if you have GPU and want memory efficiency
# xformers>=0.0.22  

# Development (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
